Add lmpwatch.pl to tests directory, and engine-side support for it in g_game.c
lmpwatch.pl is a clone of PrBoom-Plus's automated demo playback system.  Given
a demo, a suitable patterns file such as Grazza's lmpwatch.zip[1], and a list
of directories where wads are kept, it will build a command to start the game,
and execute it.

It has rudimentary support for testing demo sync.  The idea, due to Grazza[2],
is that if the player exits the map, a competn-style demo is unlikely to have
desynced.  The required engine-side support is simply that "FINISHED:
<mapname>" is printed on standard output by the game, and captured by the
script.  Unfortunately there is currently no way to check the list of finished
maps automatically, so lmpwatch.pl merely prints it out.

[1] http://www.doomworld.com/vb/showthread.php?s=&threadid=38859 , 
    http://www.doomworld.com/sda/dwdemo/lmpwatch.zip

[2] http://www.doomworld.com/vb/showthread.php?s=&postid=644382#post644382