One more 'ATI sucks' revision.
Some ATI’s drivers have a bug whereby adding the depth renderbuffer and then a texture causes the application to crash. This should be kept in mind when doing any FBO related work and tested for as it is possible it could be fixed in a future driver revision thus rendering the problem non-existent.

Some TRY/EXCEPT stuff and workarounds here